requirements.txt
requests~=2.28.1
ciso8601~=2.2.0
python-dotenv~=0.21.0
Flask~=2.2.2
gevent~=21.12.0
websocket~=0.2.1
websocket-client~=1.4.1
pandas~=1.5.0